Refactor the TableCellProperties tests.
